% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISA.r
\name{elisa_ifg}
\alias{elisa_ifg}
\title{Convert table of number of survivals per day to individual information table}
\usage{
elisa_ifg(dataset, concentration = concentration,
  standard_col = standard_col)
}
\arguments{
\item{dataset}{dataset with all the data}

\item{concentration}{concentration for each dilution}

\item{standard_col}{a vector with two values representing the standard columns}
}
\value{
a data frame with calculated concentration
}
\description{
Convert table of number of survivals per day to individual information table
}
