% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuno_survival.r
\name{raw2survival}
\alias{raw2survival}
\title{Convert table of number of survivals per day to individual information table}
\usage{
raw2survival(data = datos, num_col_group = 4)
}
\arguments{
\item{data}{dataset with survivals per day}

\item{num_col_group}{number of columns after day column}
}
\value{
a data frame with information per individual
}
\description{
Convert table of number of survivals per day to individual information table
}
